fix(navtiming): stop assigning to w.p.t.navigationStart
In 90be3c2, we tried to fix problems with negative navtiming values
by setting the fallback value for navigationStart earlier during the
initialisation phase. Mistakenly, we did this by assigning Date.now()
to window.performance.timing.navigationStart, which is a getter-only
property.

This change fixes the issue by setting a local variable instead of
the source property.